refactor app architecture
