+++ src/models/leafnode.py
@@ -1,4 +1,4 @@
-from src.htmlnode import HTMLNode
+from src.models.htmlnode import HTMLNode
 
 
 class LeafNode(HTMLNode):
