+++ src/renderer/text_to_textnodes.py
@@ -1,8 +1,8 @@
-from src.htmlnode import HTMLNode
-from src.textnode import TextNode, TextNodeType
-from src.utils.split_nodes_delimiter import split_nodes_delimiter
-from src.utils.split_nodes_image import split_nodes_image
-from src.utils.split_nodes_link import split_nodes_link
+from src.models.htmlnode import HTMLNode
+from src.models.textnode import TextNode, TextNodeType
+from src.parser.split_nodes_delimiter import split_nodes_delimiter
+from src.parser.split_nodes_image import split_nodes_image
+from src.parser.split_nodes_link import split_nodes_link
 
 
 def text_to_textnodes(text: str) -> list[TextNode | HTMLNode]:
