+++ tests/tests-models/test_htmlnode.py
@@ -1,6 +1,6 @@
 import unittest
 
-from src.htmlnode import HTMLNode
+from src.models.htmlnode import HTMLNode
 
 
 class TestHTMLNode(unittest.TestCase):
