+++ tests/tests-models/test_leafnode.py
@@ -1,6 +1,6 @@
 import unittest
 
-from src.leafnode import LeafNode
+from src.models.leafnode import LeafNode
 
 
 class TestLeafNode(unittest.TestCase):
