+++ tests/tests-parser/test_extract_markdown.py
@@ -1,6 +1,6 @@
 import unittest
 
-from src.utils.extract_markdown import extract_markdown_images, extract_markdown_links
+from src.parser.extract_markdown import extract_markdown_images, extract_markdown_links
 
 
 class TestExtractMarkdownImages(unittest.TestCase):
